1317 - add new donation page POM
