+++ e2e/data/donation-test.data.ts
@@ -1,0 +1,7 @@
+export const stripeFormData = {
+  cardNumber: '4242 4242 4242 4242',
+  email: 'e2e_test_mail@test.bg',
+  expiryDate: '04 / 42',
+  cvc: '424',
+  country: 'BG',
+}
