+++ e2e/data/support-page-tests.data.ts
@@ -5,11 +5,3 @@ export const supportPageVolutneerTestData = {
   phone: '+359888000000',
   comment: 'E2E Test comment',
 }
-
-export const anonDonationTestData = {
-  cardNumber: '4242 4242 4242 4242',
-  cardExpDate: '04 / 42',
-  cardCvc: '424',
-  billingName: 'E2E Test Anonymous Donation',
-  country: 'BG',
-}
